Fixed TWI
Sensor is working ....... :-)

+++ Regler_1/Regler_1/src/sensor.h
@@ -18,11 +18,11 @@ void sensor_init(void);
 /*
 	read_Sensor_Data liest von der Registeraddresse _addr count values aus
 */
-status_code_t read_sensor_data(bno055_register_addr_t _addr, uint_fast8_t *values, uint_fast8_t count);
+status_code_t read_sensor_data(bno055_register_addr_t _addr, uint8_t *values, uint_fast32_t count);
 
 /*
 	write_Sensor_Data schreibt an der Registeraddresse _addr count values 
 */
-status_code_t write_sensor_data(bno055_register_addr_t _addr, uint_fast8_t *values, uint_fast8_t count);
+status_code_t write_sensor_data(bno055_register_addr_t _addr, uint8_t *values, uint_fast8_t count);
 
 #endif /* SENSOR_H_ */
